get data from booking cookie

+++ src/app/pages/payment/payment-component/payment-component.component.ts
@@ -1,5 +1,20 @@
 import { Component, OnInit } from '@angular/core';
 import { CookieService } from 'ngx-cookie-service';
+import { UserService } from 'src/app/services/user.service';
+import { User } from 'src/app/models/user';
+import { BookingService } from 'src/app/services/booking.service';
+import { Booking } from 'src/app/models/booking';
+import { Room } from 'src/app/models/room';
+import { RoomService } from 'src/app/services/room.service';
+
+import { Service } from 'src/app/models/service';
+import { ServiceService } from 'src/app/services/service.service';
+
+import { PaymentService } from 'src/app/services/payment.service';
+import { Payment } from 'src/app/models/payment';
+
+import { forkJoin, Observable } from 'rxjs';
+
 
 @Component({
   selector: 'app-payment-component',
@@ -21,21 +36,84 @@ export class PaymentComponentComponent implements OnInit {
     { title: 'Tus datos', content: 'details' },
     { title: 'Resumen de compra', content: 'resumen' },
     { title: 'Pago', content: 'payment' },
-    
   ];
 
-  constructor(private cookieService: CookieService) { }
+  bookingID : string | undefined;
+  userID : string | undefined;
+  roomID : string | undefined;
 
-  ngOnInit(): void {
+  constructor(private cookieService: CookieService,
+    private bookingService : BookingService,
+    private userService : UserService,
+    private roomService: RoomService,
+    private paymentService : PaymentService) { }
+
+   ngOnInit(): void{
     this.currentContent = this.steps[0].content;
+
+    // logica de los datos
+    if (this.cookieService.check('booking_id')) {
+      const bookingID = this.cookieService.get('booking_id');
+      // obtengo la reserva a través del id de la cookie
+      this.getBooking(bookingID); //reservation
+
+    } 
+
   }
 
   ngOnDestroy(): void {
-    this.cookieService.delete('booking_id');
+    if (this.cookieService.check('booking_id')) {
+      this.cookieService.delete('booking_id');
+    }
   }
 
+  getBooking(bookingID : string):void{
 
+    try{
+      this.bookingService.getBooking(bookingID).subscribe(
+        (res) => {
+          this.paymentService.booking$ = res as Booking;
+          this.getUser(this.paymentService.booking$.userId);
+          this.getRoom(this.paymentService.booking$.roomId);
+        },
+        (err) => console.log(err)
+    
+      );
+    } catch (error) {
+      console.error(error);
+    }
+  
+  }
 
+  getUser(userID : string):void{
+      
+      try{
+        this.userService.getUser(userID).subscribe(
+          (res) => {
+            this.paymentService.user$ = res as User;
+          },
+          (err) => console.log(err)
+      
+        );
+      } catch (error) {
+        console.error(error);
+      }
+  }
+
+  getRoom(roomID : string):void{
+      
+    try{
+      this.roomService.getRoom(roomID).subscribe(
+        (res) => {
+          this.paymentService.room$ = res as Room;
+        },
+        (err) => console.log(err)
+    
+      );
+    } catch (error) {
+      console.error(error);
+    }
+  }
 
 
   previousStep() {
